refactoring folder structure and removing unnecessary files from repository
