+++ .idea/workspace.xml
@@ -3,13 +3,26 @@
   <component name="ChangeListManager">
     <list default="true" id="b265ea58-941f-4ed5-b2b2-6b5781012547" name="Default Changelist" comment="I hope that it should be working, having to add a external javafx lib">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/out/production/hackers-guide/.idea/vcs.xml" beforeDir="false" afterPath="$PROJECT_DIR$/out/production/hackers-guide/.idea/vcs.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/DoubledLinkList.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/DoubledLinkList.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/GameInfoTranslator.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/GameInfoTranslator.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/GenreTranslator.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/GenreTranslator.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/Main.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/Main.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/StartScreenController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/StartScreenController.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/Utility.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/hackers-guide/.gitignore" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/hackers-guide/.idea/.gitignore" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/hackers-guide/.idea/misc.xml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/hackers-guide/.idea/modules.xml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/hackers-guide/.idea/uiDesigner.xml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/hackers-guide/.idea/vcs.xml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/hackers-guide/hackers-guide.iml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/hackers-guide/src/gamePickerScreen.fxml" beforeDir="false" afterPath="$PROJECT_DIR$/src/view/gamePickerScreen.fxml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/hackers-guide/src/gameRecommendationScreen.fxml" beforeDir="false" afterPath="$PROJECT_DIR$/src/view/gameRecommendationScreen.fxml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/out/production/hackers-guide/src/startScreen.fxml" beforeDir="false" afterPath="$PROJECT_DIR$/src/view/startScreen.fxml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/CentralRoutingController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/controller/CentralRoutingController.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/DoubledLinkList.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/model/DoubledLinkList.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/Game.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/model/Game.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/GameInfoTranslator.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/model/GameInfoTranslator.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/GenreTranslator.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/model/GenreTranslator.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/StartScreenController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/controller/StartScreenController.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/User.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/model/User.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gamePickerScreen.fxml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/gameRecommendationScreen.fxml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/startScreen.fxml" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -50,7 +63,8 @@
     <property name="WebServerToolWindowFactoryState" value="false" />
     <property name="aspect.path.notification.shown" value="true" />
     <property name="last_opened_file_path" value="$USER_HOME$/.jdks/javafx-sdk-11.0.2/lib" />
-    <property name="project.structure.last.edited" value="Project" />
+    <property name="nodejs_package_manager_path" value="npm" />
+    <property name="project.structure.last.edited" value="Modules" />
     <property name="project.structure.proportion" value="0.15" />
     <property name="project.structure.side.proportion" value="0.2" />
     <property name="settings.editor.selected.configurable" value="reference.projectsettings.compiler.javacompiler" />
@@ -60,10 +74,11 @@
       <recent name="src" />
     </key>
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\Users\carte\IdeaProjects\hackers-guide\src\view" />
+      <recent name="C:\Users\carte\IdeaProjects\hackers-guide\src\views" />
       <recent name="C:\Users\ianal\Documents\GitHub\hackers-guide\src" />
       <recent name="C:\Users\ianal\Documents\GitHub\hackers-guide\veiws" />
       <recent name="C:\Users\ianal\Documents\GitHub\hackers-guide" />
-      <recent name="C:\Users\ianal\Documents\GitHub\hackers-guide\lib" />
     </key>
   </component>
   <component name="RunManager">
@@ -75,15 +90,13 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
-    <configuration default="true" type="ArquillianTestNG" factoryName="" nameIsGenerated="true">
+    <configuration default="true" type="ArquillianJUnit" factoryName="" nameIsGenerated="true">
       <option name="arquillianRunConfiguration">
         <value>
           <option name="containerStateName" value="" />
         </value>
       </option>
-      <option name="TEST_OBJECT" value="CLASS" />
-      <properties />
-      <listeners />
+      <option name="TEST_OBJECT" value="class" />
       <method v="2">
         <option name="Make" enabled="true" />
       </method>
@@ -108,6 +121,7 @@
       <workItem from="1617048203216" duration="45000" />
       <workItem from="1617048358133" duration="9316000" />
       <workItem from="1617115206183" duration="8578000" />
+      <workItem from="1617543253471" duration="907000" />
     </task>
     <task id="LOCAL-00001" summary="Added basic fxml beginning point">
       <created>1616684591966</created>
