+++ src/controller/CentralRoutingController.java
@@ -1,4 +1,4 @@
-package src;
+package src.controller;
 
 import javafx.scene.control.Button;
 import javafx.scene.control.CheckBox;
