+++ src/model/GenreTranslator.java
@@ -1,4 +1,4 @@
-package src;/*
+package src.model;/*
 Last updated: 24 March, 2021
 This class will call on the RAWG API and display all possible genres.
 Authors: Emily Crabtree
